+++ src/vlog/vlog-node.h
@@ -315,6 +315,7 @@ typedef enum {
    V_SYSTF_CLOG2,
    V_SYSTF_SQRT,
    V_SYSTF_BITS,
+   V_SYSTF_RTOI,
 } vlog_systf_t;
 
 typedef enum {