+++ src/vlog/vlog-sem.c
@@ -1072,6 +1072,7 @@ static type_mask_t vlog_check_sys_fcall(vlog_node_t v)
    case V_SYSTF_CLOG2:
    case V_SYSTF_SIGNED:
    case V_SYSTF_UNSIGNED:
+   case V_SYSTF_RTOI:
       return TM_INTEGRAL | (TM_CONST & mask);
    case V_SYSTF_SQRT:
       return TM_REAL | (TM_CONST & mask);