+++ test/test_vlog.c
@@ -28,6 +28,7 @@
 #include "vlog/vlog-node.h"
 #include "vlog/vlog-number.h"
 #include "vlog/vlog-phase.h"
+#include "vpi/vpi-model.h"
 
 #include <inttypes.h>
 
@@ -1430,10 +1431,12 @@ START_TEST(test_simp1)
 {
    input_from_file(TESTDIR "/vlog/simp1.v");
 
+   vpi_context_new();
+
    vlog_node_t m = do_parse_check(V_MODULE);
    vlog_simp(m);
 
-   const int64_t expect[] = { 5, 1, 7, 1, 2, 2, -4, 6, -2, -4 };
+   const int64_t expect[] = { 5, 1, 7, 1, 2, 2, -4, 6, -2, -4, 4 };
 
    ck_assert_int_eq(ARRAY_LEN(expect), vlog_decls(m));